requirements.txt
google-genai
python-dotenv
numpy
PyPDF2
groq
schedule
streamlit
schedule
requests 
pandas
